requirements.txt
#
# This file is autogenerated by zipline-pip-tools==0.0.13 with python 3.8
# See https://github.com/ZiplineTeam/zipline-pip-tools for details.
# DO NOT MODIFY BY HAND. To update, modify the .in file and run:
#
#    run_pip_compile
#
asgiref==3.7.2
    # via opentelemetry-instrumentation-asgi
backoff==2.2.1
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-http
blinker==1.7.0
    # via flask
certifi==2023.11.17
    # via requests
charset-normalizer==3.3.1
    # via requests
click==8.1.7
    # via flask
deprecated==1.2.14
    # via
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-http
flask==3.0.0
    # via
    #   -r requirements.in
    #   flask-cors
flask-cors==4.0.0
    # via -r requirements.in
googleapis-common-protos==1.61.0
    # via opentelemetry-exporter-otlp-proto-http
idna==3.4
    # via requests
importlib-metadata==6.8.0
    # via
    #   flask
    #   opentelemetry-api
itsdangerous==2.1.2
    # via flask
jinja2==3.1.3
    # via flask
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
opentelemetry-api==1.21.0
    # via
    #   -r requirements.in
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-confluent-kafka
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-flask
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-sqlalchemy
    #   opentelemetry-instrumentation-wsgi
    #   opentelemetry-sdk
opentelemetry-exporter-otlp-proto-common==1.21.0
    # via opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-http==1.21.0
    # via -r requirements.in
opentelemetry-instrumentation==0.42b0
    # via
    #   -r requirements.in
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-flask
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-sqlalchemy
    #   opentelemetry-instrumentation-wsgi
opentelemetry-instrumentation-asgi==0.42b0
    # via opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-confluent-kafka==0.43b0
    # via -r requirements.in
opentelemetry-instrumentation-fastapi==0.42b0
    # via -r requirements.in
opentelemetry-instrumentation-flask==0.42b0
    # via -r requirements.in
opentelemetry-instrumentation-requests==0.42b0
    # via -r requirements.in
opentelemetry-instrumentation-sqlalchemy==0.42b0
    # via -r requirements.in
opentelemetry-instrumentation-wsgi==0.42b0
    # via opentelemetry-instrumentation-flask
opentelemetry-proto==1.21.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.21.0
    # via
    #   -r requirements.in
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-semantic-conventions==0.42b0
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-flask
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-sqlalchemy
    #   opentelemetry-instrumentation-wsgi
    #   opentelemetry-sdk
opentelemetry-util-http==0.42b0
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-flask
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-wsgi
packaging==23.2
    # via
    #   opentelemetry-instrumentation-flask
    #   opentelemetry-instrumentation-sqlalchemy
protobuf==3.20.2
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
requests==2.31.0
    # via opentelemetry-exporter-otlp-proto-http
typing-extensions==4.8.0
    # via
    #   asgiref
    #   opentelemetry-sdk
urllib3==1.26.18
    # via requests
werkzeug==3.0.1
    # via flask
wheel==0.41.2
    # via -r requirements.in
wrapt==1.15.0
    # via
    #   deprecated
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-confluent-kafka
    #   opentelemetry-instrumentation-sqlalchemy
zipp==3.17.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
pip==23.3.1
    # via -r requirements.in
setuptools==68.2.2
    # via
    #   -r requirements.in
    #   opentelemetry-instrumentation
